pyproject.toml
[tool.poetry]
name = "fantasy-nfl"
version = "0.1.0"
description = "Fantasy NFL data pipeline"
authors = ["Branch Manager <n/a>"]
readme = "README.md"
packages = [
  { include = "fantasy_nfl", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
typer = { version = "0.9.0", extras = ["all"] }
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
pandas = "^2.2.2"
click = "^8.1.7"
playwright = "^1.46.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
fantasy = "fantasy_nfl.cli:cli"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
